Added main menu option to display readers & indexes.
Rearranged declaration, definition, initialization, and loading of some main objects:
  Reader classes, handling, naming.
  Led, RGB, BTserial.
  Hardware pins.

Blocked printing to BTserial during Menu startup.

This build runs well.

+++ settings.cpp
@@ -153,7 +153,7 @@
         break;
       case 8 :
         //sprintf(setting_value, "%s", default_reader);
-        sprintf(setting_value, "%i (%s)", default_reader, Reader::NameFromIndex(default_reader));
+        sprintf(setting_value, "%i (%s)", default_reader, Reader::NameFromIndex((int)default_reader));
         break;
       case 9 :
         sprintf(setting_value, "%li", hw_serial_baud);
